tests: Added unit test for card attachment use-cases

+++ api/src/application/use-cases/card-attachment/DeleteCardAttachmentUseCase.js
@@ -7,15 +7,26 @@ class DeleteCardAttachmentUseCase {
   }
 
   async execute(cardAttachment) {
+    if (!cardAttachment?.id) throw new Error('cardId was not provided');
+
     if (cardAttachment.type !== 'external-link') {
       const deletedInStorage = await this.cloudinaryStorageRepository.destroy(
         cardAttachment.publicId,
       );
+
       if (deletedInStorage?.result !== 'ok')
         throw Boom.badRequest('Failed to delete file from Cloudinary');
     }
 
-    return this.cardAttachmentRepository.delete(cardAttachment.id);
+    const result = await this.cardAttachmentRepository.delete(
+      cardAttachment.id,
+    );
+
+    if (result === 0) {
+      throw new Error('Something went wrong deleting the attachment');
+    }
+
+    return result;
   }
 }
